security: cap SD-JWT disclosure segments to prevent strings.Split DoS (Axis 42)
VerifySDJWTWithBinding called strings.Split(sdjwt, "~") without any limit
on the number of resulting segments. The issuer signature covers only the
first segment (the JWT header.payload.sig), leaving the rest attacker-editable.
A malicious wallet can append thousands of "~" characters to any legitimate
SD-JWT; strings.Split on a 4 MiB input of "~" allocates ~4M string-header
entries (~64 MB) before the signature check runs, enabling memory exhaustion
under concurrent load.

Fix: add package-level maxSDJWTSegments = 256 and check
strings.Count(sdjwt, "~") > maxSDJWTSegments before the split in both
VerifySDJWTWithBinding and Present. New sentinel: ErrSDJWTTooManyDisclosures.
Test: TestVerifySDJWTTooManyDisclosures covers all three entry points.

Also: remove two unnecessary uint64() casts in scitt/scitt_test.go
(golangci-lint unconvert), leaving lint at 0 issues; correct README fuzz-target
count (20→21) and test count (1000+→1700+).

+++ blrcs/scitt/scitt_test.go
@@ -640,13 +640,13 @@ func TestVerifyInclusionTamperedLastByte(t *testing.T) {
 	}
 	leaf := ledger.leafHashes[0]
 	// Correct root must verify.
-	if !VerifyInclusion(leaf, correctRoot, 0, uint64(receipt.TreeSize), path) {
+	if !VerifyInclusion(leaf, correctRoot, 0, receipt.TreeSize, path) {
 		t.Fatal("valid inclusion proof rejected")
 	}
 	// Tamper only the last byte of the root hash — must be rejected.
 	tampered := bytes.Clone(correctRoot)
 	tampered[len(tampered)-1] ^= 0xFF
-	if VerifyInclusion(leaf, tampered, 0, uint64(receipt.TreeSize), path) {
+	if VerifyInclusion(leaf, tampered, 0, receipt.TreeSize, path) {
 		t.Fatal("tampered root (last byte) should be rejected")
 	}
 }
